+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # Changelog
 
-## 8.4.0 2026-06-05
+## 9.0.0 2026-06-05
 
 * Rust
-    * Add thread-parallel stiffness matrix assembly for 2D FEM solver
+    * !Add thread-parallel stiffness matrix assembly for 2D FEM solver
+    * Add direct/iterative solve selection for 2D FEM using BiCGSTAB with optional equilibration
+    * Add optional iterative solve, equilibration, and preconditioning for 2D FEM solver
+    * Run cargo-semver-checks during testing
 * Python
+    * Add bindings plumbing for iterative solver
     * Plumb parallel option into 2D FEM interface
+    * Add 2D FEM solve diagnostics and BiCGSTAB options
 
 ## 8.3.0 2026-05-04